Add new class Ant and Positions
