+++ src/main/java/jalau/at18/katas/langtonant/Team1/Ant.java
@@ -1,0 +1,30 @@
+package jalau.at18.katas.langtonant.Team1;
+
+public class Ant {
+    private int posX;
+    private int posY;
+    
+
+    public Ant(int posX, int posY) {
+        this.posX = posX;
+        this.posY = posY;
+    }
+
+    public Ant() {
+    }
+
+
+    public int getPosX() {
+        return posX;
+    }
+    public void setPosX(int posX) {
+        this.posX = posX;
+    }
+    public int getPosY() {
+        return posY;
+    }
+    public void setPosY(int posY) {
+        this.posY = posY;
+    }
+    
+}
