+++ src/main/java/jalau/at18/katas/langtonant/Team1/Movements.java
@@ -2,12 +2,17 @@ package jalau.at18.katas.langtonant.Team1;
 
 public class Movements {
     
-    Position position = new Position();
+    Ant position = new Ant();
 
     public void moveWhite(){
+
         
     }
     public void moveBlack(){
 
+    }
+
+    public void moveRight(){
+        
     }
 }
