--- src/main/java/jalau/at18/katas/langtonant/Team1/changePosition.java
@@ -1,5 +1,0 @@
-package jalau.at18.katas.langtonant.Team1;
-
-public class changePosition {
-    
-}
